Update Program.java
Finished the whole code. Also didn't have to use the math library, to use the math methods. Asks the user for input values using the scanner method and then adds the values together, subtracts them, and multiplies them. and then finds the average, and absolute value, and then find the maximum and the minimum without using the max and min functions from the maths library.